Make it possible to complie 64-bit binary on windows
The change is to supply `x64` to MSVC_TARGETS if `-8` is specified.
Note that it will set DEST_CPU to `amd64` instead of expected `x86_64`
on Windows. This is known to break things for xash3d-fwgs repo. But for
this one it seems just fine? I haven't investigated it further.